+++ packages/confirmation/src/components/screens/hint/component.tsx
@@ -31,7 +31,7 @@ export const Hint: FC<HintProps> = ({ mobile }) => {
 
     return (
         <div className={cn(styles.component, styles[alignContent])}>
-            <Header mobile={mobile}>Не&nbsp;приходит сообщение?</Header>
+            <Header mobile={mobile}>{texts.hintTitle}</Header>
 
             <TypographyText
                 view='primary-medium'
@@ -41,8 +41,7 @@ export const Hint: FC<HintProps> = ({ mobile }) => {
                     [styles.typographyThemeMobile]: mobile,
                 })}
             >
-                Если у&nbsp;вас изменился номер телефона, позвоните нам или обратитесь в&nbsp;любое
-                отделение банка.
+                {texts.hintDescription}
             </TypographyText>
 
             <div
@@ -50,28 +49,6 @@ export const Hint: FC<HintProps> = ({ mobile }) => {
                     [styles.phonesWrapMobile]: mobile,
                 })}
             >
-                <div className={cn(styles.phoneWrap, { [styles.phoneContentMobile]: mobile })}>
-                    <Link
-                        href={getPhoneHref(domesticPhone)}
-                        underline={false}
-                        className={cn(styles.phoneLink, { [styles.typographyThemeMobile]: mobile })}
-                    >
-                        {domesticPhone}
-                    </Link>
-                    <TypographyText
-                        view='primary-medium'
-                        color='primary'
-                        className={cn(styles.text, {
-                            [styles.typographyTheme]: !mobile,
-                            [styles.typographyThemeMobile]: mobile,
-                        })}
-                    >
-                        {mobile
-                            ? 'Для\u00A0звонков по\u00A0России'
-                            : ' \u2014\u00A0для звонков по\u00A0России'}
-                    </TypographyText>
-                </div>
-
                 <div className={cn(styles.phoneWrap, { [styles.phoneContentMobile]: mobile })}>
                     <Link
                         href={getPhoneHref(internationalPhone)}
@@ -93,6 +70,28 @@ export const Hint: FC<HintProps> = ({ mobile }) => {
                             : ' \u2014\u00A0в\u00A0Москве и\u00A0за\u00A0границей'}
                     </TypographyText>
                 </div>
+
+                <div className={cn(styles.phoneWrap, { [styles.phoneContentMobile]: mobile })}>
+                    <Link
+                        href={getPhoneHref(domesticPhone)}
+                        underline={false}
+                        className={cn(styles.phoneLink, { [styles.typographyThemeMobile]: mobile })}
+                    >
+                        {domesticPhone}
+                    </Link>
+                    <TypographyText
+                        view='primary-medium'
+                        color='primary'
+                        className={cn(styles.text, {
+                            [styles.typographyTheme]: !mobile,
+                            [styles.typographyThemeMobile]: mobile,
+                        })}
+                    >
+                        {mobile
+                            ? 'Для\u00A0звонков по\u00A0России'
+                            : ' \u2014\u00A0для звонков по\u00A0России'}
+                    </TypographyText>
+                </div>
             </div>
 
             <TypographyText
@@ -103,10 +102,8 @@ export const Hint: FC<HintProps> = ({ mobile }) => {
                     [styles.typographyThemeMobile]: mobile,
                 })}
             >
-                Если номер не&nbsp;менялся, возможно, перегружен сервис отправки сообщений.
-                Попробуйте повторить действие через несколько минут.
+                {texts.hintNotification}
             </TypographyText>
-
             <Button
                 size={mobile ? 'xs' : 's'}
                 view='secondary'