+++ packages/product-cover/src/components/text/Component.tsx
@@ -9,7 +9,7 @@ import { pluralize } from '@alfalab/utils';
 
 import { TYPOGRAPHY_VIEW_FOR_SIZE } from '../../consts';
 import { Size } from '../../typings';
-import { showNumberOfСards } from '../../utils';
+import { showNumberOfCards } from '../../utils';
 
 import styles from './index.module.css';
 
@@ -45,9 +45,15 @@ export type TextProps = {
     cardNumber?: number;
 
     /**
-     * Количество карт
+     * (Устаревший) Количество карт
+     * @deprecated Используйте 'numberOfCards'
      */
     numberOfСards?: number;
+
+    /**
+     * Количество карт
+     */
+    numberOfCards?: number;
 
     /**
      * Управление ориентацией стопки карт компонента
@@ -70,6 +76,7 @@ export const Text: React.FC<TextProps> = ({
     cardholderName,
     cardNumber,
     numberOfСards,
+    numberOfCards = numberOfСards,
     align,
     eyeButton,
     dataTestId,
@@ -77,7 +84,7 @@ export const Text: React.FC<TextProps> = ({
     onEyeIconClick,
 }) => {
     const maxSize = size === 164;
-    const visibleNumberOfСards = showNumberOfСards(size, numberOfСards, align);
+    const visibleNumberOfCards = showNumberOfCards(size, numberOfCards, align);
 
     return (
         <div className={cn(styles.component)} data-test-id={dataTestId}>
@@ -93,7 +100,7 @@ export const Text: React.FC<TextProps> = ({
                 </TypographyText>
             ) : null}
 
-            {cardNumber && !visibleNumberOfСards ? (
+            {cardNumber && !visibleNumberOfCards ? (
                 <div className={styles.cardNumber}>
                     <TypographyText
                         view={TYPOGRAPHY_VIEW_FOR_SIZE[size]}
@@ -117,15 +124,15 @@ export const Text: React.FC<TextProps> = ({
                 </div>
             ) : null}
 
-            {visibleNumberOfСards && (
+            {visibleNumberOfCards && (
                 <div className={styles.cardNumber}>
                     <TypographyText
                         view={TYPOGRAPHY_VIEW_FOR_SIZE[size]}
                         weight='medium'
                         color='static-primary-dark'
                     >
-                        {`+${numberOfСards}\u00A0`}
-                        {pluralize(numberOfСards || 0, 'карта', 'карты', 'карт')}
+                        {`+${numberOfCards}\u00A0`}
+                        {pluralize(numberOfCards || 0, 'карта', 'карты', 'карт')}
                     </TypographyText>
                 </div>
             )}